refactor(#3940): split the registry's test-only barrier plumbing into registry_test_hooks.rs
registry.rs is over the 800-line source threshold, so the campsite ratchet
FAILs on growth. Move all six barrier accessors out (873 -> 848 lines, a net
shrink) and dedupe the three clone-out-of-the-lock runners onto one helper.
